+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 1.1.0 - 05.03.2021
+
+* Added `isLinux`, `isMacOS`, `isWindows`, `isDesktop`, `isFuchsia` getters
+* Removed `unversal_html` dependency
+
 ## 1.0.0 - 20.10.2020
 
-* Implemented cross platform functionality.
+* Implemented cross platform functionality